Revert OAuth2::Client behavior.
In previous versions, OAuth2 would always return with a `response` and
its corresponding status code and body. New release of OAuth2 raises
exceptions for different codes and breaks our code.

This was not found by our specs because we're mocking the response from
OAuth2 and we don't test its expected behavior.